+++ legacy_packages/react/src/wallet/wallets/embeddedWallet/embeddedWallet.tsx
@@ -123,6 +123,7 @@ export const embeddedWallet = (
     create(walletOptions: WalletOptions) {
       return new EmbeddedWallet({
         ...walletOptions,
+        ...options,
         clientId: walletOptions?.clientId ?? "",
       });
     },