+++ packages/thirdweb/src/extensions/erc1155/drops/read/getActiveClaimCondition.ts
@@ -1,4 +1,6 @@
 import type { BaseTransactionOptions } from "../../../../transaction/types.js";
+import type { ClaimCondition } from "../../../../utils/extensions/drops/types.js";
+import { claimCondition } from "../../__generated__/DropSinglePhase1155/read/claimCondition.js";
 import {
   type GetActiveClaimConditionIdParams,
   getActiveClaimConditionId,
@@ -20,7 +22,43 @@ export type GetActiveClaimConditionParams = GetActiveClaimConditionIdParams;
  */
 export async function getActiveClaimCondition(
   options: BaseTransactionOptions<GetActiveClaimConditionParams>,
-) {
-  const conditionId = await getActiveClaimConditionId(options);
-  return getClaimConditionById({ ...options, conditionId });
+): Promise<ClaimCondition> {
+  const getActiveClaimConditionMultiPhase = async () => {
+    const conditionId = await getActiveClaimConditionId(options);
+    return getClaimConditionById({ ...options, conditionId });
+  };
+  const getActiveClaimConditionSinglePhase = async () => {
+    const [
+      startTimestamp,
+      maxClaimableSupply,
+      supplyClaimed,
+      quantityLimitPerWallet,
+      merkleRoot,
+      pricePerToken,
+      currency,
+      metadata,
+    ] = await claimCondition({ ...options, tokenId: options.tokenId });
+    return {
+      startTimestamp,
+      maxClaimableSupply,
+      supplyClaimed,
+      quantityLimitPerWallet,
+      merkleRoot,
+      pricePerToken,
+      currency,
+      metadata,
+    };
+  };
+
+  // The contract's phase type is unknown, so try both options and return whichever resolves, prioritizing multi-phase
+  const results = await Promise.allSettled([
+    getActiveClaimConditionMultiPhase(),
+    getActiveClaimConditionSinglePhase(),
+  ]);
+
+  const condition = results.find((result) => result.status === "fulfilled");
+  if (condition?.status === "fulfilled") {
+    return condition.value;
+  }
+  throw new Error("Claim condition not found");
 }